requirements.txt
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
vaderSentiment>=3.3.0
requests>=2.25.0
google-genai>=0.2.0
Flask>=2.0.0
gunicorn>=20.1.0
streamlit>=1.30.0
pandas>=2.0.0
